+++ Lecture_code/model_wideanddeep.py
@@ -1,0 +1,20 @@
+
+from tensorflow.keras.layers import Input, Dense, Concatenate
+from tensorflow.keras import Model
+
+class WideAndDeepModel(Model):
+    def __init__(self, units=30, activation="relu", **kwargs):
+        super().__init__(**kwargs)
+        self.hidden1 = Dense(units, activation=activation)
+        self.hidden2 = Dense(units, activation=activation)
+        self.main_output = Dense(1)
+        self.aux_output = Dense(1)
+
+    def call(self, inputs):
+        input_a, input_b = inputs
+        hidden1 = self.hidden1(input_b)
+        hidden2 = self.hidden2(hidden1)
+        concat = Concatenate()([input_a, input_b])
+        main_output = self.main_output(concat)
+        aux_output = self.aux_output(hidden2)
+        return main_output, aux_output
